Add route queue observability
Add tracing and optional metrics to the Apalis route queue
adapter so enqueue failures are diagnosable in production.

- instrumented enqueue and provider setup/push paths with tracing,
  including warning logs on serialization and enqueue failures
- added optional Prometheus enqueue counters and latency histograms
- extracted metrics wiring into a dedicated module and routed to a
  no-op fallback when feature metrics is disabled
- moved Apalis queue tests to a dedicated module for clarity
- updated queue architecture docs and kept Generic/Apalis references
  consistent
