+++ backend/src/outbound/queue/apalis_route_queue/tests.rs
@@ -1,0 +1,162 @@
+//! Unit tests for the Apalis route queue adapter.
+
+use super::*;
+use crate::outbound::queue::test_helpers::{FailingQueueProvider, FakeQueueProvider};
+use rstest::rstest;
+use serde::{Deserialize, Serialize};
+
+/// Test plan type for unit tests.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+struct TestPlan {
+    name: String,
+}
+
+#[rstest]
+#[tokio::test]
+async fn apalis_queue_enqueue_round_trips() {
+    let fake_provider = FakeQueueProvider::new();
+    let queue: GenericApalisRouteQueue<TestPlan, _> =
+        GenericApalisRouteQueue::new(fake_provider.clone());
+
+    let plan = TestPlan {
+        name: "test-plan".to_string(),
+    };
+
+    let result = queue.enqueue(&plan).await;
+    assert!(result.is_ok(), "enqueue should succeed with fake provider");
+
+    let pushed_jobs = fake_provider
+        .pushed_jobs()
+        .expect("should be able to access pushed jobs");
+    assert_eq!(pushed_jobs.len(), 1, "exactly one job should be pushed");
+
+    // Verify the payload can be deserialized back to the original plan
+    let deserialized: TestPlan = serde_json::from_value(pushed_jobs[0].clone())
+        .expect("pushed payload should be valid JSON");
+    assert_eq!(
+        deserialized, plan,
+        "deserialized plan should match original"
+    );
+}
+
+#[rstest]
+#[tokio::test]
+async fn apalis_queue_maps_provider_error_to_unavailable() {
+    let failing_provider = FailingQueueProvider::new("simulated queue failure".to_string());
+    let queue: GenericApalisRouteQueue<TestPlan, _> =
+        GenericApalisRouteQueue::new(failing_provider);
+
+    let plan = TestPlan {
+        name: "test-plan".to_string(),
+    };
+
+    let result = queue.enqueue(&plan).await;
+    assert!(
+        result.is_err(),
+        "enqueue should fail when provider returns error"
+    );
+
+    match result.expect_err("expected error but call succeeded") {
+        JobDispatchError::Unavailable { message } => {
+            assert!(
+                message.contains("simulated queue failure"),
+                "error message should contain provider error: {message}"
+            );
+        }
+        JobDispatchError::Rejected { .. } => {
+            panic!("expected Unavailable error, got Rejected");
+        }
+    }
+}
+
+#[rstest]
+#[tokio::test]
+async fn apalis_queue_enqueues_multiple_plans() {
+    let fake_provider = FakeQueueProvider::new();
+    let queue: GenericApalisRouteQueue<TestPlan, _> =
+        GenericApalisRouteQueue::new(fake_provider.clone());
+
+    let plan1 = TestPlan {
+        name: "plan-1".to_string(),
+    };
+    let plan2 = TestPlan {
+        name: "plan-2".to_string(),
+    };
+
+    queue
+        .enqueue(&plan1)
+        .await
+        .expect("first enqueue should succeed");
+    queue
+        .enqueue(&plan2)
+        .await
+        .expect("second enqueue should succeed");
+
+    let pushed_jobs = fake_provider
+        .pushed_jobs()
+        .expect("should be able to access pushed jobs");
+    assert_eq!(pushed_jobs.len(), 2, "both jobs should be pushed");
+
+    let deserialized1: TestPlan =
+        serde_json::from_value(pushed_jobs[0].clone()).expect("first payload should be valid JSON");
+    let deserialized2: TestPlan = serde_json::from_value(pushed_jobs[1].clone())
+        .expect("second payload should be valid JSON");
+
+    assert_eq!(deserialized1, plan1, "first plan should match");
+    assert_eq!(deserialized2, plan2, "second plan should match");
+}
+
+/// Test plan type that always fails serialization.
+#[derive(Debug, Clone)]
+struct FailingSerializePlan;
+
+impl Serialize for FailingSerializePlan {
+    fn serialize<S>(&self, _serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        Err(serde::ser::Error::custom("simulated serialization failure"))
+    }
+}
+
+#[rstest]
+#[tokio::test]
+async fn apalis_queue_maps_serialization_failure_to_rejected() {
+    let fake_provider = FakeQueueProvider::new();
+    let queue: GenericApalisRouteQueue<FailingSerializePlan, _> =
+        GenericApalisRouteQueue::new(fake_provider.clone());
+
+    let plan = FailingSerializePlan;
+
+    let result = queue.enqueue(&plan).await;
+    assert!(
+        result.is_err(),
+        "enqueue should fail when serialization fails"
+    );
+
+    match result.expect_err("expected error but call succeeded") {
+        JobDispatchError::Rejected { message } => {
+            assert!(
+                message.contains("Failed to serialize plan"),
+                "error message should contain adapter context: {message}"
+            );
+            assert!(
+                message.contains("simulated serialization failure"),
+                "error message should contain underlying serializer error: {message}"
+            );
+        }
+        JobDispatchError::Unavailable { .. } => {
+            panic!("expected Rejected error for serialization failure, got Unavailable");
+        }
+    }
+
+    // Verify nothing was pushed to the provider
+    let pushed_jobs = fake_provider
+        .pushed_jobs()
+        .expect("should be able to access pushed jobs");
+    assert_eq!(
+        pushed_jobs.len(),
+        0,
+        "no jobs should be pushed when serialization fails"
+    );
+}
